+++ src/sZIP.App/AuditWindow.xaml.cs
@@ -1,5 +1,3 @@
-using System.Diagnostics;
-using System.IO;
 using System.Windows;
 
 namespace sZIP.App;
@@ -18,22 +16,6 @@ public partial class AuditWindow : Window
     private void RefreshButton_Click(object sender, RoutedEventArgs e) => Refresh();
 
     private void Localization_Changed(object? sender, EventArgs e) => Refresh();
-
-    private void OpenFolderButton_Click(object sender, RoutedEventArgs e)
-    {
-        try
-        {
-            Directory.CreateDirectory(DiagnosticLog.LogDirectory);
-            Process.Start(new ProcessStartInfo("explorer.exe", DiagnosticLog.LogDirectory)
-            {
-                UseShellExecute = true
-            });
-        }
-        catch (Exception exception)
-        {
-            DiagnosticLog.Write("audit-folder.open.failed", exception);
-        }
-    }
 
     private void Refresh() =>
         AuditGrid.ItemsSource = AutomaticArchiveExtractionAudit.ReadRecent();